Stop closing HandshakeConductor's udpConn in its private methods
The super caller should be responsible for this. If any of
HandshakeConductor's exported methods return an error, the udpConn
should be prematurely closed when it's handled.

This makes things more readable, as well: whatever is responsible for
initializing udpConn and passing it into a HandshakeConductor's
constructor should also be responsible for closing that udpConn
prematurely when appropriate.