Set I/O timeouts on all _net.Conn & _net.PacketConn vars
This serves as a rudimentary way to avoid hanging indefinitely if a
connection between a sender and receiver is dropped, or if a sender
never reaches out to a waiting receiver to begin the lancp handshake.

The error messages that these timeouts generate aren't friendly or
explanatory enough to display to users. That'll need to be addressed
Soon (TM).